# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Guest Service API",
        "description": "API documentation for Guest Service",
        "contact": {
            "name": "Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/guests": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "List all guests",
                "operationId": "ba13faba9c30d75ce7386db1cee72e22",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Guest"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Guests"
                ],
                "summary": "Create a new guest",
                "operationId": "abf1a99cd422a8b8b54c4a4ff63461f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Guest created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/guests/{id}": {
            "get": {
                "tags": [
                    "Guests"
                ],
                "summary": "Get a single guest",
                "operationId": "cc61bee00307fa8ce7d4041a328e1b54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Guests"
                ],
                "summary": "Update an existing guest",
                "operationId": "23fcd53cd995340d5018b111a6f722fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Guest updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Guest"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Guests"
                ],
                "summary": "Delete a guest",
                "operationId": "40fd2e7cb98f2852adbc8d7c4d5c8ee6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Guest ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Guest deleted"
                    },
                    "404": {
                        "description": "Guest not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Guest": {
                "title": "Guest",
                "description": "A guest record",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "GuestInput": {
                "title": "GuestInput",
                "description": "Input data for creating or updating a guest",
                "properties": {
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "country": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Guests",
            "description": "Guests"
        }
    ]
}